register per app top nav items (#53136)
