+++ test/plugin_functional/plugins/kbn_tp_top_nav/index.js
@@ -25,6 +25,7 @@ export default function(kibana) {
         description: 'This is a sample plugin for the functional tests.',
         main: 'plugins/kbn_tp_top_nav/app',
       },
+      hacks: ['plugins/kbn_tp_top_nav/initialize'],
     },
   });
 }
